app.tsx 복구
