+++ FrontEnd/meet-mate/mobile-content/promise/promiseHome.tsx
@@ -1,0 +1,9 @@
+
+function PromiseHome() {
+    return (
+    <div>
+
+    </div>
+    );
+}
+export default PromiseHome;
