--- FrontEnd/meet-mate/mobile/promise/promiseHome.tsx
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-const PromisDiv = styled.div`
-    
-`;
-
-function PromiseHome() {
-    return (
-    <PromisDiv>
-        gg
-    </PromisDiv>
-    );
-}
-export default PromiseHome;
