+++ FrontEnd/meet-mate/pages/mobile/index.tsx
@@ -3,12 +3,11 @@ import Image from 'next/image'
 import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 import { Map,MapMarker } from 'react-kakao-maps-sdk'
-import { IMarkers, mapAtom } from '@/mobile/atom'
+import { IMarkers, mapAtom } from '@/mobile-content/atom'
 import { useRecoilValue } from 'recoil'
 const inter = Inter({ subsets: ['latin'] })
 
 export default function Home(props:any) {
-  console.log(props);
   const mapRecoil = useRecoilValue<IMarkers>(mapAtom);
   return (
     <>
